+++ src/main/java/asia/asoulcnki/api/persistence/entity/UserDynamic.java
@@ -28,5 +28,5 @@ public class UserDynamic implements Serializable {
 	private int status;
 
     @JsonProperty("user_id")
-    private long userID;
+    private int userID;
 }
